# openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Script Labs API",
    "description": "A secure Script Labs API built with Node.js and Express, featuring JWT authentication via Supabase and CRUD operations for labs management.",
    "version": "1.0.0",
    "contact": {
      "name": "Hendri Christianto",
      "email": "hendri.christianto24@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://script-labs-app-z8p8.onrender.com",
      "description": "Production server on Render"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/labs": {
      "get": {
        "summary": "Get all labs",
        "description": "Retrieve all labs belonging to the authenticated user",
        "tags": ["Labs"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of labs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lab"
                  }
                },
                "example": [
                  {
                    "id": 14,
                    "title": "buku menari",
                    "author": "jojo",
                    "user_id": "cd004999d-7d4b-4167-a9af-0f15783caee5"
                  },
                  {
                    "id": 15,
                    "title": "buku mewarnai",
                    "author": "santi",
                    "user_id": "cd004999d-7d4b-4167-a9af-0f15783caee5"
                  }
                ]
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "summary": "Add a new lab",
        "description": "Create a new lab entry for the authenticated user",
        "tags": ["Labs"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookInput"
              },
              "example": {
                "title": "Atomic Habits",
                "author": "James Clear"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "lab created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/lab"
                },
                "example": {
                  "id": 25,
                  "title": "Atomic Habits",
                  "author": "James Clear",
                  "user_id": "cd004999d-7d4b-4167-a9af-0f15783caee5"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Title and author are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from Supabase authentication"
      }
    },
    "schemas": {
      "lab": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the lab",
            "example": 14
          },
          "title": {
            "type": "string",
            "description": "Title of the lab",
            "example": "Atomic Habits"
          },
          "author": {
            "type": "string",
            "description": "Author of the lab",
            "example": "James Clear"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user who owns this lab",
            "example": "cd004999d-7d4b-4167-a9af-0f15783caee5"
          }
        },
        "required": ["id", "title", "author", "user_id"]
      },
      "BookInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the lab",
            "example": "Atomic Habits"
          },
          "author": {
            "type": "string",
            "description": "Author of the lab",
            "example": "James Clear"
          }
        },
        "required": ["title", "author"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "An error occurred"
          },
          "error": {
            "type": "string",
            "description": "Detailed error information"
          }
        },
        "required": ["message"]
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication information is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "no_token": {
                "summary": "No token provided",
                "value": {
                  "message": "No token provided"
                }
              },
              "invalid_token": {
                "summary": "Invalid token",
                "value": {
                  "message": "Invalid token"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "Server error saat menambah buku",
              "error": "Database connection failed"
            }
          }
        }
      }
    }
  },
...
}